Redirect to game after login
