+++ react/tic-tac-toe/src/app/index.js
@@ -1,8 +1,7 @@
 import React from 'react';
 import { Route, Switch, Redirect } from 'react-router-dom';
-import { connect } from 'react-redux';
 
-import AuthorizatedRoute from './screens/Login/components/authorizatedRoute';
+import AuthorizedRoute from './screens/Login/components/authorizedRoute';
 import Game from './screens/Game';
 import Login from './screens/Login'
 
@@ -12,15 +11,10 @@ function App() {
   return (
     <Switch>
       <Route path="/login" component={Login} />
-      <AuthorizatedRoute path="/game" component={Game} />
+      <AuthorizedRoute path="/game" component={Game} />
       <Redirect to="/login" />
     </Switch> 
   )
 }
 
-const mapStateToProps = state => ({
-  isLogged: state.login.isLogged,
-  token: state.login.token
-});
-
-export default connect(mapStateToProps)(App);
+export default App;
